requirements.txt
Pillow>=10.0.0
watchdog>=3.0.0
